Added default return values for calls where address is undefined
The address of a tile will not be defined for coordinates out of bounds, like Point(-1,-1); because there exists no c-structure holding any data on this imaginary tile.

Added check to see if coordinates of a tile is within x[0,7] y[0,7] before searching for its address; defaulting to NULL for coordinates out of bounds.